Move CLI unit tests into an integration target
Keep the lib harness under Clippy's large_stack_arrays ceiling after
merging main, which reintroduced enough registrations to exceed 2048.
